INitialize secret and arg_from_vault. Defensive check when deleting secret

+++ src/mist/api/clouds/controllers/main/base.py
@@ -291,6 +291,9 @@ class BaseMainController(object):
                 'msg': "Invalid parameters %s." % list(errors.keys()),
                 'errors': errors,
             })
+
+        secret = None
+        arg_from_vault = False
         # Set fields to cloud model and perform early validation.
         for key, value in kwargs.items():
             if value:  # ignore empty values
@@ -349,7 +352,7 @@ class BaseMainController(object):
             log.error("Error updating %s: %s", self.cloud, exc.to_dict())
             # delete VaultSecret object and secret
             # if it was just added to Vault
-            if not arg_from_vault:
+            if not arg_from_vault and secret:
                 secret.delete()
                 secret.ctl.delete_secret()
             raise BadRequestError({'msg': str(exc),
